requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.4
httpx>=0.27.0
jinja2>=3.1.4
pyyaml>=6.0.0
sqlalchemy>=2.0.32
python-multipart>=0.0.9
typing_extensions>=4.7.0
